// libs/admin-plugins/reports/daily-order/src/generated-admin-types.ts
export type Maybe<T> = T | null;

/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  JSON: any;
  Upload: any;
  Date: any;
};
// ...
export type DailyOrderInput = {
  month?: Maybe<Scalars['String']>;
  year: Scalars['String'];
  state: Scalars['String'];
};

export type DailyOrderList = PaginatedList & {
   __typename?: 'DailyOrderList';
  items: Array<Order>;
  totalItems: Scalars['Int'];
};

export type DailyOrderOutput = {
   __typename?: 'DailyOrderOutput';
  items: Array<OrdersPerMonth>;
};
// ...
export type OrdersPerMonth = {
   __typename?: 'OrdersPerMonth';
  month: Scalars['String'];
  totalOrders: Scalars['Int'];
  day?: Maybe<Scalars['String']>;
};
// ...
export type QueryDailyOrdersArgs = {
  input: DailyOrderInput;
};